requirements.txt
dDjango>=5.0.0
openpyxl
pandas
filelock
mysqlclient>=2.2.0
numpy
scipy
statsmodels
nltk
python-dotenv
xlsxwriter
matplotlib
django-allauth>=0.57.0
seaborn
apscheduler
marshmallow
tablib
watchdog
xlrd
xlwt
